Adding a registration handler for jenkins.
THis is for test runs on Jenkins to push results to a test plan
on testrail. This is for creating a brand new testrun under a test
plan and then pushing any results that RSpect happens to give you.

That is the opposite of running test cases pre-defined on testrail.

Change-Id: I296fad8f2923e66e0c0add9786f2dcaae54eaf3f