update run once status of testcase
this commit makes two small changes. the primary
change marks any testcase found in spec/requests/
as run once. the secondary change was a quality
of life change. now we print out more of the
title of each testcase and end it with an
ellipsis if it is too long to display

Change-Id: If93d7678adc6662ec41439c6c1effa9dcfea5810