update testcases marked as automated without automation
this commit adds an additional check while updating
automation status. for any test not found in the
spec files, it will make sure that they are not
marked as automated or have a spec location. if they
do, we make sure that the previous code is working
appropriately before updating the testcase to
reflect the codebase

Change-Id: If5aef3bf20d20eb8992910ea07ce8a5ef2daaa17